Implement UnloadPlugin()
For implementation the program need to remove physical plugin's assembly from plugin directory.
When the assembly load into AppDomain the file is in use by ClientUI process and can not be removed and there is no way to unload an specific assembly from AppDomain but we can unload the entire AppDomain which is not usefull for the problem.
For fix this issue i use a trick and load the plugin assembly from memory instead of directly from physical file .